third_party/blink/web_tests/platform/mac/virtual/threaded-composited-iframes/external/wpt/is-input-pending/tentative/same-origin-subframe.sub-expected.txt

This is a testharness.js-based test.
FAIL can detect same-origin iframe events in parent assert_equals: detected discrete input expected true but got false
PASS can detect same-origin iframe events in subframe
Harness: the test ran to completion.
